chore: finished up waypoint screen rework
