+++ common/src/main/java/dev/ftb/mods/ftbchunks/client/gui/LargeMapScreen.java
@@ -138,13 +138,7 @@ public class LargeMapScreen extends BaseScreen {
 				.append(FTBChunksClient.INSTANCE.waypointManagerKey.getTranslatedKeyMessage())
 				.append(Component.literal("]")).withStyle(ChatFormatting.GRAY);
 		add(waypointManagerButton = new SimpleTooltipButton(this, Component.translatable("ftbchunks.gui.waypoints"), Icons.COMPASS,
-				(b, m) -> {
-			if(isShiftKeyDown()) {
-				new WaypointEditorScreen().openGui();
-			}else {
-				new UpdatedEditorScreen().openGui();
-			}
-                }, tooltip));
+				(b, m) -> new WaypointEditorScreen().openGui(), tooltip));
 		add(infoButton = new SimpleButton(this, Component.translatable("ftbchunks.gui.large_map_info"), Icons.INFO,
 				(b, m) -> new MapKeyReferenceScreen().openGui()));
