+++ common/src/main/java/dev/ftb/mods/ftbchunks/client/gui/WaypointEditorScreen.java
@@ -1,6 +1,5 @@
 package dev.ftb.mods.ftbchunks.client.gui;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import dev.architectury.networking.NetworkManager;
 import dev.ftb.mods.ftbchunks.client.map.MapManager;
 import dev.ftb.mods.ftbchunks.client.map.WaypointImpl;
@@ -10,213 +9,176 @@ import dev.ftb.mods.ftblibrary.config.StringConfig;
 import dev.ftb.mods.ftblibrary.icon.Color4I;
 import dev.ftb.mods.ftblibrary.icon.Icon;
 import dev.ftb.mods.ftblibrary.icon.Icons;
-import dev.ftb.mods.ftblibrary.icon.ImageIcon;
 import dev.ftb.mods.ftblibrary.ui.*;
 import dev.ftb.mods.ftblibrary.ui.input.Key;
 import dev.ftb.mods.ftblibrary.ui.input.MouseButton;
+import dev.ftb.mods.ftblibrary.ui.misc.AbstractButtonListScreen;
 import dev.ftb.mods.ftblibrary.util.TooltipList;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.FormattedText;
+import net.minecraft.network.chat.HoverEvent;
 import net.minecraft.network.chat.Style;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
 import org.lwjgl.glfw.GLFW;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
-public class WaypointEditorScreen extends BaseScreen {
-    public static final Color4I COLOR_BACKGROUND = Color4I.rgba(0xC0333333);
-    private static final Icon PEARL_ICON = ImageIcon.getIcon(ResourceLocation.fromNamespaceAndPath("minecraft", "textures/item/ender_pearl.png"));
+import static dev.ftb.mods.ftblibrary.util.TextComponentUtils.hotkeyTooltip;
+import static net.minecraft.network.chat.CommonComponents.ELLIPSIS;
 
-    public static final Theme THEME = new Theme() {
-        @Override
-        public void drawScrollBarBackground(GuiGraphics graphics, int x, int y, int w, int h, WidgetType type) {
-            Color4I.BLACK.withAlpha(70).draw(graphics, x, y, w, h);
-        }
-
-        @Override
-        public void drawScrollBar(GuiGraphics graphics, int x, int y, int w, int h, WidgetType type, boolean vertical) {
-            getContentColor(WidgetType.NORMAL).withAlpha(100).withBorder(Color4I.GRAY.withAlpha(100), false).draw(graphics, x, y, w, h);
-        }
-    };
-
-    private static final Component TITLE = Component.translatable("ftbchunks.gui.waypoints").withStyle(ChatFormatting.BOLD);
-
-    private final Panel waypointPanel;
-    private final PanelScrollBar scrollbar;
-    private final List<Widget> rows;
+public class WaypointEditorScreen extends AbstractButtonListScreen {
+    private final Map<ResourceKey<Level>, Boolean> collapsed = new HashMap<>();
+    private final Map<ResourceKey<Level>, List<WaypointImpl>> waypoints = new HashMap<>();
     private final Button buttonCollapseAll, buttonExpandAll;
-    private int maxWidth;
 
     public WaypointEditorScreen() {
-        waypointPanel = new WaypointPanel(this);
-        scrollbar = new PanelScrollBar(this, waypointPanel) {
-            @Override
-            public void drawBackground(GuiGraphics graphics, Theme theme, int x, int y, int w, int h) {
-                if (getMaxValue() > parent.height) {
-                    super.drawBackground(graphics, theme, x, y, w, h);
-                }
-            }
-        };
-        rows = new ArrayList<>();
+        showBottomPanel(false);
+        showCloseButton(true);
 
-        buttonCollapseAll = new ExpanderButton(this, false);
-        buttonExpandAll = new ExpanderButton(this, true);
-
-        for (DimWayPoints entry : collectWaypoints()) {
-            rows.add(new VerticalSpaceWidget(waypointPanel,4));
-
-            boolean thisDimension = Minecraft.getInstance().level.dimension().location().equals(entry.levelKey);
-            GroupButton groupButton = new GroupButton(waypointPanel, Component.literal(entry.levelKey.toString()).withStyle(ChatFormatting.YELLOW), !thisDimension);
-            rows.add(groupButton);
-
-            for (WaypointImpl wp : entry.waypoints) {
-                rows.add(new RowPanel(waypointPanel, groupButton, wp));
-            }
+        for (Map.Entry<ResourceKey<Level>, List<WaypointImpl>> resourceKeyListEntry : collectWaypoints().entrySet()) {
+            collapsed.put(resourceKeyListEntry.getKey(), false);
+            waypoints.put(resourceKeyListEntry.getKey(), new ArrayList<>(resourceKeyListEntry.getValue()));
         }
+
+        buttonExpandAll = new SimpleButton(topPanel, List.of(Component.translatable("gui.expand_all"), hotkeyTooltip("="), hotkeyTooltip("+")), Icons.UP,
+                (widget, button) -> toggleAll(false));
+        buttonCollapseAll = new SimpleButton(topPanel, List.of(Component.translatable("gui.collapse_all"), hotkeyTooltip("-")), Icons.DOWN,
+                (widget, button) -> toggleAll(true));
     }
 
-    private List<DimWayPoints> collectWaypoints() {
-        List<DimWayPoints> res = new ArrayList<>();
+    private static FormattedText ellipsize(Font font, FormattedText text, int maxWidth) {
+        final int strWidth = font.width(text);
+        final int ellipsisWidth = font.width(ELLIPSIS);
+        if (strWidth > maxWidth) {
+            return ellipsisWidth >= maxWidth ?
+                    font.substrByWidth(text, maxWidth) :
+                    FormattedText.composite(font.substrByWidth(text, maxWidth - ellipsisWidth), ELLIPSIS);
+        }
+        return text;
+    }
 
-        Player player = Objects.requireNonNull(Minecraft.getInstance().player);
-        MapManager manager = MapManager.getInstance().orElseThrow();
-        manager.getDimensions().values().stream()
-                .filter(dim -> !dim.getWaypointManager().isEmpty())
-                .sorted((dim1, dim2) -> {
-                    // put vanilla dimensions first
-                    ResourceLocation dim1id = dim1.dimension.location();
-                    ResourceLocation dim2id = dim2.dimension.location();
-                    if (dim1id.getNamespace().equals("minecraft") && !dim2id.getNamespace().equals("minecraft")) {
-                        return -1;
-                    }
-                    int i = dim1id.getNamespace().compareTo(dim2id.getNamespace());
-                    return i == 0 ? dim1id.getPath().compareTo(dim2id.getPath()) : i;
-                })
-                .forEach(dim -> res.add(new DimWayPoints(dim.dimension.location(), dim.getWaypointManager().stream()
-                        .sorted(Comparator.comparingDouble(wp -> wp.getDistanceSq(player)))
-                        .toList()))
-                );
+    private void toggleAll(boolean collapsed) {
+        this.collapsed.keySet().forEach(levelResourceKey -> this.collapsed.put(levelResourceKey, collapsed));
+        scrollBar.setValue(0);
+        getGui().refreshWidgets();
+    }
 
-        return res;
+    @Override
+    protected void doCancel() {
+    }
+
+    @Override
+    protected void doAccept() {
     }
 
     @Override
     public boolean onInit() {
-        boolean r = setFullscreen();
-        maxWidth = Math.min(maxWidth, getGui().width / 2);
-        return r;
+        setWidth(220);
+        setHeight(getScreen().getGuiScaledHeight() * 4 / 5);
+        return true;
     }
 
     @Override
-    public void addWidgets() {
-        add(waypointPanel);
-        add(scrollbar);
-
-        add(buttonCollapseAll);
-        add(buttonExpandAll);
+    protected int getTopPanelHeight() {
+        return 22;
     }
 
     @Override
-    public void alignWidgets() {
-        waypointPanel.setPosAndSize(0, 20, width, height - 20);
-        waypointPanel.alignWidgets();
-        scrollbar.setPosAndSize(width - 16, 20, 16, height - 20);
-        buttonExpandAll.setPos(width - 18, 2);
-        buttonCollapseAll.setPos(width - 38, 2);
+    protected Panel createTopPanel() {
+        return new CustomTopPanel();
     }
 
     @Override
-    public Theme getTheme() {
-        return super.getTheme();
+    public void addButtons(Panel panel) {
+        waypoints.forEach((key, value) -> {
+            boolean startCollapsed = collapsed.get(key);
+            GroupButton groupButton = new GroupButton(panel, key, startCollapsed, value);
+            panel.add(groupButton);
+            if (!startCollapsed) {
+                panel.addAll(groupButton.collectPanels());
+            }
+        });
     }
 
-    @Override
-    public Component getTitle() {
-        return TITLE;
-    }
-
-    @Override
-    public void drawBackground(GuiGraphics graphics, Theme theme, int x, int y, int w, int h) {
-        Color4I.BLACK.withAlpha(128).draw(graphics, 0, 0, w, h);
-        COLOR_BACKGROUND.draw(graphics, 0, 0, w, 20);
-        theme.drawString(graphics, getTitle(), 6, 6, Theme.SHADOW);
-    }
-
-    @Override
-    public boolean keyPressed(Key key) {
-        if (key.esc() && getContextMenu().isPresent()) {
-            closeContextMenu();
-            return true;
-        }
-        return super.keyPressed(key);
-    }
-
-    private class WaypointPanel extends Panel {
-        public WaypointPanel(Panel panel) {
-            super(panel);
-        }
+    protected class CustomTopPanel extends TopPanel {
+        private final TextField titleLabel = new TextField(this);
 
         @Override
         public void addWidgets() {
-            maxWidth = 0;
-            for (var row : rows) {
-                if (!(row instanceof RowPanel rowPanel) || !rowPanel.groupButton.collapsed && !rowPanel.deleted) {
-                    add(row);
-                    if (row instanceof RowPanel rp) maxWidth = Math.max(maxWidth, getTheme().getStringWidth(rp.wp.getName()));
-                }
+            titleLabel.setText(Component.translatable("ftbchunks.gui.waypoints"));
+            titleLabel.addFlags(Theme.CENTERED_V);
+            add(titleLabel);
+
+            if (waypoints.size() > 1) {
+                add(buttonExpandAll);
+                add(buttonCollapseAll);
             }
-            maxWidth = Math.min(maxWidth, getGui().width / 2);
         }
 
         @Override
         public void alignWidgets() {
-            align(WidgetLayout.VERTICAL);
-            int sbWidth = scrollbar.getMaxValue() > 0 ? 16 : 0;
-            scrollbar.setWidth(sbWidth);
-            widgets.forEach(w -> {
-                int prevWidth = w.width;
-                w.setWidth(width - sbWidth);
-                if (w.width != prevWidth && w instanceof Panel p) p.alignWidgets();
-            });
+            titleLabel.setPosAndSize(4, 0, titleLabel.width, height);
+            if (waypoints.size() > 1) {
+                buttonExpandAll.setPos(width - 18, 2);
+                buttonCollapseAll.setPos(width - 38, 2);
+            }
         }
     }
 
-    private static class GroupButton extends Button {
-        private boolean collapsed;
+    private class GroupButton extends Button {
         private final Component titleText;
+        private final List<RowPanel> rowPanels;
+        private final ResourceKey<Level> dim;
 
-        public GroupButton(Panel panel, Component titleText, boolean startCollapsed) {
+        public GroupButton(Panel panel, ResourceKey<Level> dim, boolean startCollapsed, List<WaypointImpl> waypoints) {
             super(panel);
-            this.titleText = titleText;
+
+            this.dim = dim;
+            //Todo move the to library or define / get some standard for dim names
+            //Might need to show more with dim types / level ids
+            this.titleText = Component.translatableWithFallback(dim.location().toLanguageKey("dimension"), dim.location().getPath())
+                    .withStyle(style -> style.withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal(dim.location().toString()))));
             setCollapsed(startCollapsed);
+            this.rowPanels = new ArrayList<>();
+            for (WaypointImpl waypoint : waypoints) {
+                rowPanels.add(new RowPanel(panel, waypoint));
+            }
+        }
+
+        public List<RowPanel> collectPanels() {
+            return isCollapsed() ? List.of() : List.copyOf(rowPanels);
         }
 
         @Override
         public void onClicked(MouseButton button) {
-            setCollapsed(!collapsed);
-            getGui().refreshWidgets();
+            setCollapsed(!isCollapsed());
+            parent.refreshWidgets();
+            refreshWidgets();
+            playClickSound();
         }
 
-        void setCollapsed(boolean collapsed) {
-            this.collapsed = collapsed;
-            setTitle(Component.literal(collapsed ? "[>] " : "[v] ").withStyle(collapsed ? ChatFormatting.RED : ChatFormatting.GREEN).append(titleText));
+        public boolean isCollapsed() {
+            return collapsed.get(dim);
+        }
+
+        public void setCollapsed(boolean collapsed) {
+            WaypointEditorScreen.this.collapsed.put(dim, collapsed);
+            boolean isCollapsed = isCollapsed();
+            setTitle(Component.literal(isCollapsed ? "▶ " : "▼ ").withStyle(isCollapsed ? ChatFormatting.RED : ChatFormatting.GREEN).append(titleText));
         }
 
         @Override
         public void draw(GuiGraphics graphics, Theme theme, int x, int y, int w, int h) {
-            COLOR_BACKGROUND.draw(graphics, x, y, w, h);
-            theme.drawString(graphics, getTitle(), x + 3, y + 1 + (h - theme.getFontHeight()) / 2);
-            RenderSystem.setShaderColor(1F, 1F, 1F, 1F);
-
-            Color4I.GRAY.withAlpha(80).draw(graphics, 0, y, width, 1);
-            Color4I.GRAY.withAlpha(80).draw(graphics, 0, y, 1, height);
+            theme.drawWidget(graphics, x, y, w, h, getWidgetType());
+            theme.drawString(graphics, getTitle(), x + 3, y + 3);
             if (isMouseOver()) {
                 Color4I.WHITE.withAlpha(33).draw(graphics, x, y, w, h);
             }
@@ -228,55 +190,64 @@ public class WaypointEditorScreen extends BaseScreen {
     }
 
     private class RowPanel extends Panel {
+        private static final Component DELETE = Component.translatable("ftbchunks.gui.delete");
+        private static final Component QUICK_DELETE = Component.translatable("ftbchunks.gui.quick_delete");
 
-        private static final Component DELETE = Component.literal("Delete");
-        private static final Component QUICK_DELETE = Component.literal("Quick Delete");
-
-        private final GroupButton groupButton;
         private final WaypointImpl wp;
-        private boolean deleted = false;
+        private TextField nameField;
+        private TextField distField;
+        private SimpleButton hideButton;
+        private SimpleButton deleteButton;
 
-        public RowPanel(Panel panel, GroupButton groupButton, WaypointImpl wp) {
+        public RowPanel(Panel panel, WaypointImpl wp) {
             super(panel);
-            this.groupButton = groupButton;
+
             this.wp = wp;
             setHeight(18);
         }
 
         @Override
         public void addWidgets() {
-            add(new SimpleButton(this, Component.empty(), wp.isHidden() ? Icons.REMOVE_GRAY : Icons.ACCEPT, (w, mb) -> {
+            add(hideButton = new SimpleButton(this, Component.empty(), wp.isHidden() ? Icons.REMOVE_GRAY : Icons.ACCEPT, (w, mb) -> {
                 wp.setHidden(!wp.isHidden());
                 w.setIcon(wp.isHidden() ? Icons.REMOVE_GRAY : Icons.ACCEPT);
             }));
 
-            add(new TextField(this).setMaxWidth(maxWidth).setTrim().setText(wp.getName()).setColor(Color4I.rgb(wp.getColor())).addFlags(Theme.SHADOW));
+            add(nameField = new TextField(this).setTrim().setColor(Color4I.rgb(wp.getColor())).addFlags(Theme.SHADOW));
 
             LocalPlayer player = Minecraft.getInstance().player;
             String distStr = player.level().dimension().equals(wp.getDimension()) ?
                     String.format("%.1fm", Math.sqrt(wp.getDistanceSq(player))) : "";
-            add(new TextField(this).setText(distStr).setColor(Color4I.WHITE));
-            add(new SimpleButton(this, DELETE, Icons.BIN, (w, mb) -> deleteWaypoint(!isShiftKeyDown())) {
-
+            add(distField = new TextField(this).setText(distStr).setColor(Color4I.WHITE));
+            add(deleteButton = new SimpleButton(this, DELETE, Icons.BIN, (w, mb) -> deleteWaypoint(!isShiftKeyDown())) {
                 @Override
                 public Component getTitle() {
-                    if(isShiftKeyDown()) {
-                        return QUICK_DELETE;
-                    }else {
-                        return DELETE;
-                    }
+                    return isShiftKeyDown() ? QUICK_DELETE : DELETE;
                 }
             });
         }
 
         @Override
         public void alignWidgets() {
-            if (widgets.size() == 4) {
+        }
+
+        @Override
+        public void setWidth(int newWidth) {
+            super.setWidth(newWidth);
+
+            if (newWidth > 0) {
+                int farRight = newWidth - 8;
+
                 int yOff = (this.height - getTheme().getFontHeight()) / 2 + 1;
-                widgets.get(0).setPos(10, 1);
-                widgets.get(1).setPos(30, yOff);
-                widgets.get(2).setPos(maxWidth + 40, yOff);
-                widgets.get(3).setPos(maxWidth + 80, 1);
+
+                hideButton.setPos(farRight - 8 - 16, 1);
+                deleteButton.setPos(farRight - 8, 1);
+
+                distField.setPos(hideButton.getPosX() - 5 - distField.width, yOff);
+
+                nameField.setPos(5, yOff);
+                nameField.setText(ellipsize(getTheme().getFont(), Component.literal(wp.getName()),distField.getPosX() - 5).getString());
+                nameField.setHeight(getTheme().getFontHeight() + 2);
             }
         }
 
@@ -325,7 +296,7 @@ public class WaypointEditorScreen extends BaseScreen {
                     }));
                 }
                 if (Minecraft.getInstance().player.hasPermissions(2)) {  // permissions are checked again on server!
-                    list.add(new ContextMenuItem(Component.translatable("ftbchunks.gui.teleport"), PEARL_ICON, btn -> {
+                    list.add(new ContextMenuItem(Component.translatable("ftbchunks.gui.teleport"), Icons.ART, btn -> {
                         NetworkManager.sendToServer(new TeleportFromMapPacket(wp.getPos().above(), false, wp.getDimension()));
                         closeGui(false);
                     }));
@@ -340,25 +311,25 @@ public class WaypointEditorScreen extends BaseScreen {
 
         @Override
         public boolean keyPressed(Key key) {
-            if(key.is(GLFW.GLFW_KEY_DELETE)) {
+            if (key.is(GLFW.GLFW_KEY_DELETE)) {
                 deleteWaypoint(!isShiftKeyDown());
                 return true;
-            }else {
+            } else {
                 return super.keyPressed(key);
             }
         }
 
         private void deleteWaypoint(boolean gui) {
-            if(gui) {
+            if (gui) {
                 getGui().openYesNo(Component.translatable("ftbchunks.gui.delete_waypoint", Component.literal(wp.getName())
                         .withStyle(Style.EMPTY.withColor(wp.getColor()))), Component.empty(), () -> {
                     wp.removeFromManager();
-                    deleted = true;
+                    WaypointEditorScreen.this.waypoints.get(wp.getDimension()).remove(wp);
                     getGui().refreshWidgets();
                 });
-            }else {
+            } else {
                 wp.removeFromManager();
-                deleted = true;
+                WaypointEditorScreen.this.waypoints.get(wp.getDimension()).remove(wp);
                 getGui().refreshWidgets();
             }
         }
@@ -373,16 +344,28 @@ public class WaypointEditorScreen extends BaseScreen {
         }
     }
 
-    private class ExpanderButton extends SimpleButton {
-        public ExpanderButton(Panel panel, boolean expand) {
-            super(panel, Component.translatable(expand ? "gui.expand_all" : "gui.collapse_all"), expand ? Icons.DOWN : Icons.UP, (btn, mb) -> {
-                rows.stream().filter(w -> w instanceof GroupButton).map(w -> (GroupButton) w).forEach(g -> g.setCollapsed(!expand));
-                scrollbar.setValue(0);
-                btn.getGui().refreshWidgets();
-            });
-        }
-    }
+    private static Map<ResourceKey<Level>, List<WaypointImpl>> collectWaypoints() {
+        Map<ResourceKey<Level>, List<WaypointImpl>> res = new HashMap<>();
 
-    private record DimWayPoints(ResourceLocation levelKey, List<WaypointImpl> waypoints) {
+        Player player = Objects.requireNonNull(Minecraft.getInstance().player);
+        MapManager manager = MapManager.getInstance().orElseThrow();
+        manager.getDimensions().values().stream()
+                .filter(dim -> !dim.getWaypointManager().isEmpty())
+                .sorted((dim1, dim2) -> {
+                    // put vanilla dimensions first
+                    ResourceLocation dim1id = dim1.dimension.location();
+                    ResourceLocation dim2id = dim2.dimension.location();
+                    if (dim1id.getNamespace().equals("minecraft") && !dim2id.getNamespace().equals("minecraft")) {
+                        return -1;
+                    }
+                    int i = dim1id.getNamespace().compareTo(dim2id.getNamespace());
+                    return i == 0 ? dim1id.getPath().compareTo(dim2id.getPath()) : i;
+                })
+                .forEach(dim -> res.put(dim.dimension, dim.getWaypointManager().stream()
+                        .sorted(Comparator.comparingDouble(wp -> wp.getDistanceSq(player)))
+                        .toList())
+                );
+
+        return res;
     }
 }
