+++ common/src/main/resources/assets/ftbchunks/lang/en_us.json
@@ -115,6 +115,8 @@
 	"ftbchunks.gui.large_map_info.text": "Mouse\nLeft Button;Click/Drag to move map view\nRight Button;Context Menu\nMouse Wheel;Rotate to zoom\n\nKeys\nSpace;Center view on player\nC;Open chunk claim screen\nT;Teleport to point (op required)\nS;Open settings screen\nCtrl + S;Open server settings screen (op required)\nF3+G;Toggle Chunk Grid",
 	"ftbchunks.gui.chunk_info": "Chunk Claiming Reference",
 	"ftbchunks.gui.chunk_info.text": "Claiming\nLeft Button;Drag to claim an area\nRight Button;Drag to unclaim an area\n\nForceloading\nShift + Left Button;Drag to forceload an area\nShift + Right Button;Drag to unforceload an area\nMouse Wheel;Rotate on forceloaded chunk to adjust auto-expiry\n\nMisc\nTab;Hold to hide chunk grid\nAlt;Hold to show absolute chunk claim/force times",
+	"ftbchunks.gui.delete": "Delete",
+	"ftbchunks.gui.quick_delete": "Quick Delete",
 	"ftbteamsconfig.ftbchunks": "FTB Chunks Properties",
 	"ftbteamsconfig.ftbchunks.allow_fake_players": "Allow All Fake Players",
 	"ftbteamsconfig.ftbchunks.allow_fake_players.tooltip": "Treat ALL fake players as allies of the team.\nWARNING: Setting this to true could allow hostile players to access your claims via any fake player. Set this to false if you're unsure.",
